+++ README.md
@@ -1,11 +1,11 @@
 # city-explorer-api
 
-
 **Author**: bashar taamneh
-**Version**: 1.0.0 (increment the patch/fix version number if you make more commits past your first submission)
+**Version**: 1.1.0
 
 ## Overview
 <!-- Provide a high level overview of what this application is and why you are building it, beyond the fact that it's an assignment for this class. (i.e. What's your problem domain?) -->
+Refactoring into Modules
 
 ## Getting Started
 <!-- What are the steps that a user must take in order to build this app on their own machine and get it running? -->
@@ -14,6 +14,7 @@
 <!-- Provide a detailed description of the application design. What technologies (languages, libraries, etc) you're using, and any other relevant design information. -->
 
 ## Change Log
+
 <!-- Use this area to document the iterative changes made to your application as each feature is successfully implemented. Use time stamps. Here's an example:
 
 01-01-2001 4:59pm - Application now has a fully-functional express server, with a GET route for the location resource. -->
@@ -21,15 +22,14 @@
 ## Credit and Collaborations
 <!-- Give credit (and a link) to other people or resources that helped you build this application. -->
 
-
 ~~~
 ~~~
-Name of feature: ________________________________
 
-Estimate of time needed to complete: _____
+Name of feature: Refactoring into Modules
 
-Start time: _____
+Estimate of time needed to complete: 15 minutes
 
-Finish time: _____
+Start time: 10:15 am
+Finish time: 10:35 am
 
-Actual time needed to complete: _____
+Actual time needed to complete: 15 minutes
